templates/python-web-template/requirements.txt
fastapi
uvicorn
pydantic
requests
PyJWT[crypto]
